ENH: move pointTransformer architecture to saxi_nets and saxi_layers
